Conditional tests: examples to play with
The tests in this commit aren't "real" tests, as in: you shouldn't be testing native PHP functions in your own test suite.

These tests are just here to demonstrate the different effects of various ways to conditionally skip a test.

See the documentation of each test for more details.

To see the different effects:
- On PHP 7.3+, run `phpunit --filter skiptest` and all tests will be run and pass.
- On PHP < 7.3, run `phpunit --filter skiptest`.
    - Depending on the PHPUnit version being used, the test being skipped without skip marking will be marked as risky (no assertions).
- On PHP < 7.3, run `phpunit --filter skiptest -v` and see the skipped tests being listed.
    - Take note of only 4 out of 6 skip notices showing an explanation of *why* they are being skipped.